==> Proyecto/Versiones Previas/Originales/00_control.ipynb <==
from datetime import datetime
import os

NOMBRE_EXPERIMENTO = 'Clase6_PredJulio2018'

FILES = [
    ['01_sell_in_group.parquet',
        '01_prod_stats.parquet',
        '01_prod_data.parquet',
        '01_stock_data.parquet'],
    ['02_sell_in_norm.parquet'],
    ['03_sell_in_lag.parquet'],
    ['04_sell_in_train.parquet',
        '04_sell_in_test.parquet',
        '04_sell_in_futuro.parquet'],
    ['05_sell_in_pred.parquet',
        '05_sell_in_pred_futuro.parquet'],
    ['06_kaggle_pred.csv'],
]

FOLDER = f'./Experimentos/{NOMBRE_EXPERIMENTO}/'

for FILE in FILES:
    for f in FILE:
        print(f"{FOLDER}{f}")

%run 01_LecturaDatos.ipynb
%run 02_normaliza.ipynb
%run 03_features.ipynb
%run 04_crear_train_set.ipynb
%run 05_lightgbm.ipynb
%run 06_evaluar_prediccion.ipynb
