==> Proyecto/Versiones Previas/Originales/01_LecturaDatos.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
import os

try:
    FILES
    FILE = FILES[0]
    FOLDER
except NameError:
    FILE = []
    FILE.append('01_sell_in_group.parquet')
    FILE.append('01_prod_stats.parquet')
    FILE.append('01_prod_data.parquet')
    FILE.append('01_stock_data.parquet')
    FOLDER = 'DefaultFolder'


print(f"------------------------------COMIENZA-----------------------------\n[{FOLDER}/{FILE}")

df = pd.read_csv('../sell-in.txt',sep='\t')
prod_a_predecir = pd.read_csv('../productos_a_predecir.txt', sep='\t')
prod_data = pd.read_csv('../tb_productos.txt', sep='\t')
stock_data = pd.read_csv('../tb_stocks.txt', sep='\t')

for col in prod_data.select_dtypes(include=['object']).columns:
    prod_data[col] = prod_data[col].astype('category')

prod_data = prod_data.drop_duplicates(subset='product_id')

df_grouped = df.groupby(['periodo', 'product_id']).sum().reset_index()

df_grouped = df_grouped[['product_id', 'periodo', 'tn']]
df_grouped.sort_values(by=['product_id', 'periodo'], inplace=True)
df_grouped.reset_index(drop=True, inplace=True)

print(df_grouped.shape)
df_grouped.head()

prod_stats = df_grouped.groupby('product_id').apply(lambda group: pd.Series({
    'primer_periodo': group['periodo'].min(),
    'ultimo_periodo': group['periodo'].max(),
    'values': group['tn'].count(),
    'total_tn': group['tn'].sum(),
    'min_tn': group['tn'].min(),
    'average_tn': group['tn'].mean(),
    'median_tn': group['tn'].median(),
    'std_dev_tn': group['tn'].std(),
    'iqr_tn': (group['tn'].quantile(0.75) - group['tn'].quantile(0.25)),
    'max_tn': group['tn'].max()
}))
prod_stats['primer_periodo'] = prod_stats['primer_periodo'].astype(int)
prod_stats['primer_periodo'] = pd.to_datetime(prod_stats['primer_periodo'].astype(str), format='%Y%m')

prod_stats['ultimo_periodo'] = prod_stats['ultimo_periodo'].astype(int)
prod_stats['ultimo_periodo'] = pd.to_datetime(prod_stats['ultimo_periodo'].astype(str), format='%Y%m')

prod_stats.reset_index(inplace=True)

prod_stats = prod_stats[prod_stats['product_id'].isin(prod_a_predecir['product_id'])]
df_grouped = df_grouped[df_grouped['product_id'].isin(prod_a_predecir['product_id'])]

#prod_stats = prod_stats
#df_grouped = df_grouped

print(prod_stats.shape)
prod_stats.head()

if not os.path.exists(FOLDER):
    os.makedirs(FOLDER)

df_grouped.to_parquet(f'{FOLDER}/{FILE[0]}', index=False)
prod_stats.to_parquet(f'{FOLDER}/{FILE[1]}', index=False)
prod_data.to_parquet(f'{FOLDER}/{FILE[2]}', index=False)
stock_data.to_parquet(f'{FOLDER}/{FILE[3]}', index=False)

import matplotlib.pyplot as plt

dibujo = pd.DataFrame(df_grouped['tn'].groupby(df_grouped['periodo']).aggregate('mean')).reset_index()
dibujo['fecha'] = pd.to_datetime(dibujo['periodo'].astype(str), format='%Y%m')
dibujo.head()

# plt.figure(figsize=(20, 7))
# plt.plot(dibujo['fecha'], dibujo['tn'], marker='o')
# plt.title('Total TN by Time')
# plt.xlabel('Time')
# plt.ylabel('Total TN')
# plt.grid(True)
# plt.tight_layout()

print(f" {FOLDER}/{FILE}\n------------------------------FINALIZA-----------------------------")
