==> Proyecto/Versiones Previas/Originales/02_normaliza.ipynb <==
import pandas as pd
from datetime import datetime
import os

try:
    FILES
    FILE = FILES[1]
    FOLDER
except NameError:
    FILE = ['02_sell_in_norm.parquet']
    FOLDER = 'DefaultFolder'

    
print(f"------------------------------COMIENZA-----------------------------\n[{FOLDER}/{FILE}")

#df = pd.read_parquet(f"sell_in_group.parquet")
#prod_stats = pd.read_parquet('prod_stats.parquet')

df = pd.read_parquet(f"{FOLDER}/{FILES[0][0]}")
prod_stats = pd.read_parquet(f'{FOLDER}/{FILES[0][1]}')

df_norm = df.merge(prod_stats, how='left', on='product_id')
df_norm.head()

df_norm['tn_norm'] = (df_norm['tn'] - df_norm['average_tn']) / df_norm['std_dev_tn']

df_norm[['product_id', 'periodo','tn_norm']]

df_norm.to_parquet(f'{FOLDER}/{FILE[0]}', index=False)

df_norm[df_norm['periodo'] == 201912].shape

#estado_control = f"02_normaliza Terminado - {nombrefile} - {datetime.now()}"

print(f" {FOLDER}/{FILE}\n------------------------------FINALIZA-----------------------------")
