==> Proyecto/Versiones Previas/Originales/04_crear_train_set.ipynb <==
import pandas as pd
from datetime import datetime

import warnings

warnings.filterwarnings('ignore')

try:
    FILES
    FILE = FILES[3]
    FOLDER
except NameError:
    FILE = ['04_sell_in_train.parquet',
            '04_sell_in_test.parquet',
            '04_sell_in_futuro.parquet']
    FOLDER = 'DefaultFolder'

    
print(f"------------------------------COMIENZA-----------------------------\n[{FOLDER}/{FILE}")

periodo_train_start = 201701
periodo_train_end = 201807
periodo_test = 201807
periodo_futuro = 201912

#df = pd.read_parquet("sell_in_lag.parquet")
df = pd.read_parquet(f"{FOLDER}/{FILES[2][0]}")
print(df.shape)

df.drop(columns=['tn','values','total_tn','average_tn','median_tn','std_dev_tn','iqr_tn','max_tn','min_tn'], inplace=True)

df['primer_periodo'] = pd.to_datetime(df['primer_periodo']).dt.strftime('%Y%m').astype(int)
df['ultimo_periodo'] = pd.to_datetime(df['ultimo_periodo']).dt.strftime('%Y%m').astype(int)
df['periodo_dt'] = pd.to_datetime(df['periodo_dt']).dt.strftime('%Y%m').astype(int)
df['tn_futuro'] = -df.groupby('product_id')['tn_norm'].diff(-2)

df_train = df[(df['periodo']>=periodo_train_start) & (df['periodo']<=periodo_train_end)]
df_test  = df[df['periodo']==periodo_test]
df_futuro  = df[df['periodo']==periodo_futuro]

print(f"df_train: {df_train.shape}")
print(f"df_test: {df_test.shape}")
print(f"df_futuro: {df_test.shape}")

print(f"Shape Train: {df_train.shape} para los periodos desde {periodo_train_start} hasta {periodo_train_end}")
print(f"Shape Test: {df_test.shape} para el periodo {periodo_test}")
print(f"Shape Futuro: {df_futuro.shape} para el periodo {periodo_futuro}")

# df_train.to_parquet('sell_in_train.parquet', index=False)
# df_test.to_parquet('sell_in_test.parquet', index=False)
# df_futuro.to_parquet('sell_in_futuro.parquet', index=False)

df_train.to_parquet(f'{FOLDER}/{FILE[0]}', index=False)
df_test.to_parquet(f'{FOLDER}/{FILE[1]}', index=False)
df_futuro.to_parquet(f'{FOLDER}/{FILE[2]}', index=False)

#estado_control = f"04_crear_train_set Terminado - {nombrefile} - {datetime.now()}"

print(f" {FOLDER}/{FILE}\n------------------------------FINALIZA-----------------------------")
