==> Proyecto/Versiones Previas/Originales/05_lightgbm.ipynb <==
import pandas as pd
import lightgbm as lgb
from datetime import datetime
#from lightgbm import LGBMRegressor

try:
    FILES
    FILE = FILES[4]
    FOLDER
except NameError:
    FILE = ['05_sell_in_pred.parquet',
            '05_sell_in_pred_futuro.parquet']
    FOLDER = 'DefaultFolder'


print(f"------------------------------COMIENZA-----------------------------\n[{FOLDER}/{FILE}")

# df_train = pd.read_parquet('sell_in_train.parquet')
# df_test = pd.read_parquet('sell_in_test.parquet')
# df_futuro = pd.read_parquet('sell_in_futuro.parquet')
df_train = pd.read_parquet(f"{FOLDER}/{FILES[3][0]}")
df_test = pd.read_parquet(f"{FOLDER}/{FILES[3][1]}")
df_futuro = pd.read_parquet(f"{FOLDER}/{FILES[3][2]}")

print(f"df_train: {df_train.shape}")
print(f"df_test: {df_test.shape}")
print(f"df_futuro: {df_futuro.shape}")

categorical_features = df_train.select_dtypes(['category']).columns.tolist()
for col in categorical_features:
    print(f"convertida {col}")
    df_train[col] = df_train[col].cat.codes
    df_test[col] = df_test[col].cat.codes
    df_futuro[col] = df_futuro[col].cat.codes

X_train = df_train.iloc[:,:-1]
X_test = df_test.iloc[:,:-1]
X_futuro = df_futuro.iloc[:,:-1]

y_train = df_train.iloc[:,-1]
y_test = df_test.iloc[:,-1]
y_futuro = df_futuro.iloc[:,-1]

print(f"X_train: {X_train.shape}")
print(f"X_test: {X_test.shape}")
print(f"X_futuro: {X_futuro.shape}")

print(f"y_train: {y_train.shape}")
print(f"y_test: {y_test.shape}")
print(f"y_futuro: {y_futuro.shape}")

train_data = lgb.Dataset(X_train, label=y_train)
test_data = lgb.Dataset(X_test, label=y_test)
futuro_data = lgb.Dataset(X_futuro, label=y_futuro)

params = {
    'objective': 'regression',
    'metric': {'l2', 'rmse'},
    'boosting_type': 'gbdt',
    'num_leaves': 40,
#    'max_depth': -1,
    'learning_rate': 0.02,
#    'feature_fraction': 0.4
}

model = lgb.train(params, train_data, num_boost_round=1000, valid_sets=[train_data, test_data])

#y_pred = model.predict(X_test, num_iteration=model.best_iteration)
y_pred = model.predict(X_test, num_iteration=model.best_iteration)
y_pred_futuro = model.predict(X_futuro, num_iteration=model.best_iteration)

#example_row = X_test.iloc[0]
#example_pred = model.predict(example_row.values.reshape(1, -1), num_iteration=model.best_iteration)
#print(f'Prediction for example row: {example_pred[0]}')

final = X_test
final['tn_futuro'] = y_test
final['tn_prediccion'] = y_pred

#prod_stats = pd.read_parquet('prod_stats.parquet')
prod_stats = pd.read_parquet(f"{FOLDER}/{FILES[0][1]}")
prod_stats = prod_stats[['product_id', 'average_tn','std_dev_tn']]
final = final.merge(prod_stats, how='left', on='product_id')
final.columns

final['tn_futuro_real'] = (final['tn_norm'] + final['tn_futuro']) * final['std_dev_tn'] + final['average_tn']
final['tn_prediccion_real'] = (final['tn_norm'] + final['tn_prediccion']) * final['std_dev_tn'] + final['average_tn']

final.head()

#final.to_parquet('sell_in_pred.parquet', index=False)
final.to_parquet(f'{FOLDER}/{FILE[0]}', index=False)

final_futuro = X_futuro
final_futuro['tn_futuro'] = y_futuro
final_futuro['tn_prediccion'] = y_pred_futuro

final_futuro = final_futuro.merge(prod_stats, how='left', on='product_id')
final_futuro.columns

final_futuro['tn_futuro_real'] = (final_futuro['tn_norm'] + final_futuro['tn_futuro']) * final_futuro['std_dev_tn'] + final_futuro['average_tn']
final_futuro['tn_prediccion_real'] = (final_futuro['tn_norm'] + final_futuro['tn_prediccion']) * final_futuro['std_dev_tn'] + final_futuro['average_tn']

#final_futuro.to_parquet('sell_in_pred_futuro.parquet', index=False)
final_futuro.to_parquet(f'{FOLDER}/{FILE[1]}', index=False)

#estado_control = f"05_lightgbm Terminado - {nombrefile} - {datetime.now()}"

print(f" {FOLDER}/{FILE}\n------------------------------FINALIZA-----------------------------")
