==> Proyecto/Versiones Previas/Originales/06_evaluar_prediccion.ipynb <==
import pandas as pd
from datetime import datetime

import warnings
warnings.filterwarnings("ignore")

try:
    FILES
    FILE = FILES[5]
    FOLDER
except NameError:
    FILE = ['06_sell_in_pred_futuro.parquet',
            '06_kaggle_pred.csv']
    FOLDER = 'DefaultFolder'


print(f"------------------------------COMIENZA-----------------------------\n[{FOLDER}/{FILE}")

#df = pd.read_parquet('sell_in_pred.parquet')
df = pd.read_parquet(f"{FOLDER}/{FILES[4][0]}")
df.head()

#df_fut = pd.read_parquet('sell_in_pred_futuro.parquet')
df_fut = pd.read_parquet(f"{FOLDER}/{FILES[4][1]}")
df_fut.head()

def perdida(real, prediccion):
    total = real.sum()
    perdida =  abs(real - prediccion).sum()
    resultado = perdida / total
    return resultado

df_pred = df[['product_id', 'tn_futuro_real', 'tn_prediccion_real']]
df_pred['error'] = -(df_pred['tn_futuro_real'] - df_pred['tn_prediccion_real'])/df_pred['tn_futuro_real']*100
df_pred['weight'] = df_pred['tn_futuro_real']/df_pred['tn_futuro_real'].sum()
df_pred

perdida = perdida(df_pred['tn_futuro_real'], df_pred['tn_prediccion_real'])
perdida

# df_fut = df_fut[['product_id', 'tn_futuro_real', 'tn_prediccion_real']]
# df_fut =df_fut.merge(df_pred, on='product_id', how='left', suffixes=['FUTURO', 'VALIDACION'])
# df_fut['tn_prediccion_real'] = df_fut['tn_prediccion_realFUTURO'] * (1 -df_fut['error']/100)
# df_fut

kaggle =  df_fut[['product_id','tn_prediccion_real']]
kaggle['tn_prediccion_real'] = kaggle['tn_prediccion_real'].clip(lower=0)
kaggle.columns = ['product_id', 'tn']
#kaggle.to_csv('kaggle-new.csv', index=False)
kaggle.to_csv(f'{FOLDER}/{FILE[0]}', index=False)


#estado_control = f"06_evaluar_prediccion Terminado - {nombrefile} - {datetime.now()}"

print(f" {FOLDER}/{FILE}\n------------------------------FINALIZA-----------------------------")
